# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["user"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["user"]}},"/api/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["user"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["user"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["user"]}},"/api/user/username/{username}":{"get":{"operationId":"UserController_findOneByName","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"]}},"/api/user/{userId}/session":{"post":{"operationId":"UserSessionController_create","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserSessionDto"}}}},"responses":{"201":{"description":""}},"tags":["User/Session"]},"get":{"operationId":"UserSessionController_findAllAsPlayer","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Session"]}},"/api/user/{userId}/session/mySession":{"get":{"operationId":"UserSessionController_findAllAsMaster","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Session"]}},"/api/user/{userId}/session/{id}/mySession":{"get":{"operationId":"UserSessionController_findOneAsMaster","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Session"]}},"/api/user/{userId}/session/{id}":{"get":{"operationId":"UserSessionController_findOneAsPlayer","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Session"]},"patch":{"operationId":"UserSessionController_update","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserSessionDto"}}}},"responses":{"200":{"description":""}},"tags":["User/Session"]},"delete":{"operationId":"UserSessionController_delete","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Session"]}},"/api/user/{userId}/session/{id}/join":{"post":{"operationId":"UserSessionController_join","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["User/Session"]}},"/api/user/{userId}/session/{id}/leave":{"post":{"operationId":"UserSessionController_leave","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["User/Session"]}},"/api/user/{userId}/session/{id}/status/{status}":{"post":{"operationId":"UserSessionController_changeStatus","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"status","required":true,"in":"path","schema":{"enum":["ONLINE","OFFLINE","COMPLETED","CLOSED"],"type":"string"}}],"responses":{"201":{"description":""}},"tags":["User/Session"]}},"/api/user/{userId}/sheet":{"post":{"operationId":"UserSheetController_create","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserSheetDto"}}}},"responses":{"201":{"description":""}},"tags":["User/Sheet"]},"get":{"operationId":"UserSheetController_findAll","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Sheet"]}},"/api/user/{userId}/sheet/{id}":{"get":{"operationId":"UserSheetController_findOne","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Sheet"]},"patch":{"operationId":"UserSheetController_update","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserSheetDto"}}}},"responses":{"200":{"description":""}},"tags":["User/Sheet"]},"delete":{"operationId":"UserSheetController_delete","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Sheet"]}},"/api/user/{userId}/session/{sessionId}/sheet":{"post":{"operationId":"UserSessionSheetController_create","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"sessionId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserSessionSheetDto"}}}},"responses":{"201":{"description":""}},"tags":["User/Session/Sheet"]},"get":{"operationId":"UserSessionSheetController_findAll","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"sessionId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Session/Sheet"]}},"/api/user/{userId}/session/{sessionId}/sheet/{id}":{"get":{"operationId":"UserSessionSheetController_findOne","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"sessionId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Session/Sheet"]},"patch":{"operationId":"UserSessionSheetController_update","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"sessionId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserSessionSheetDto"}}}},"responses":{"200":{"description":""}},"tags":["User/Session/Sheet"]},"delete":{"operationId":"UserSessionSheetController_remove","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}},{"name":"sessionId","required":true,"in":"path","schema":{"type":"number"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User/Session/Sheet"]}},"/api/session":{"post":{"operationId":"SessionController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSessionDto"}}}},"responses":{"201":{"description":"The record has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Session"}}}}},"tags":["Session"]},"get":{"operationId":"SessionController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Session"]}},"/api/session/{id}":{"get":{"operationId":"SessionController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Session"]},"patch":{"operationId":"SessionController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSessionDto"}}}},"responses":{"200":{"description":""}},"tags":["Session"]},"delete":{"operationId":"SessionController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Session"]}},"/api/sheet":{"post":{"operationId":"SheetController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSheetDto"}}}},"responses":{"201":{"description":""}},"tags":["Sheet"]},"get":{"operationId":"SheetController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Sheet"]}},"/api/sheet/{id}":{"get":{"operationId":"SheetController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Sheet"]},"patch":{"operationId":"SheetController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSheetDto"}}}},"responses":{"200":{"description":""}},"tags":["Sheet"]},"delete":{"operationId":"SheetController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Sheet"]}},"/api/roll":{"post":{"operationId":"RollController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRollDto"}}}},"responses":{"201":{"description":""}},"tags":["Roll"]},"get":{"operationId":"RollController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Roll"]}},"/api/roll/{id}":{"get":{"operationId":"RollController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Roll"]},"patch":{"operationId":"RollController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRollDto"}}}},"responses":{"200":{"description":""}},"tags":["Roll"]},"delete":{"operationId":"RollController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Roll"]}}},"info":{"title":"RPG API","description":"Swagger API Documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"username":{"type":"string"},"avatar":{"type":"string"},"type":{"type":"string","enum":["MASTER","PLAYER"]}},"required":["name","username","avatar","type"]},"User":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"username":{"type":"string"},"avatar":{"type":"string"},"type":{"type":"string"}},"required":["id","name","username","avatar","type"]},"UpdateUserDto":{"type":"object","properties":{}},"CreateUserSessionDto":{"type":"object","properties":{"name":{"type":"string"},"background":{"type":"string"},"avatar":{"type":"string"}},"required":["name"]},"UpdateUserSessionDto":{"type":"object","properties":{"name":{"type":"string"},"background":{"type":"string"},"avatar":{"type":"string"}}},"CreateUserSheetDto":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string"},"image":{"type":"string"},"alignment":{"type":"string"},"background":{"type":"string"},"role":{"type":"string"},"level":{"type":"number"},"experience":{"type":"number"},"health":{"type":"number"},"strength":{"type":"number"},"dexterity":{"type":"number"},"intelligence":{"type":"number"},"constitution":{"type":"number"},"sessionId":{"type":"number"}},"required":["name","gender","image","alignment","background","role","level","experience","health","strength","dexterity","intelligence","constitution","sessionId"]},"UpdateUserSheetDto":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string"},"image":{"type":"string"},"alignment":{"type":"string"},"background":{"type":"string"},"role":{"type":"string"},"level":{"type":"number"},"experience":{"type":"number"},"health":{"type":"number"},"strength":{"type":"number"},"dexterity":{"type":"number"},"intelligence":{"type":"number"},"constitution":{"type":"number"},"sessionId":{"type":"number"}}},"CreateUserSessionSheetDto":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string"},"image":{"type":"string"},"alignment":{"type":"string"},"background":{"type":"string"},"role":{"type":"string"},"level":{"type":"number"},"experience":{"type":"number"},"health":{"type":"number"},"strength":{"type":"number"},"dexterity":{"type":"number"},"intelligence":{"type":"number"},"constitution":{"type":"number"}},"required":["name","gender","image","alignment","background","role","level","experience","health","strength","dexterity","intelligence","constitution"]},"UpdateUserSessionSheetDto":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string"},"image":{"type":"string"},"alignment":{"type":"string"},"background":{"type":"string"},"role":{"type":"string"},"level":{"type":"number"},"experience":{"type":"number"},"health":{"type":"number"},"strength":{"type":"number"},"dexterity":{"type":"number"},"intelligence":{"type":"number"},"constitution":{"type":"number"}}},"CreateSessionDto":{"type":"object","properties":{"name":{"type":"string"},"background":{"type":"string"},"masterId":{"type":"number"},"avatar":{"type":"string"}},"required":["name","masterId"]},"Session":{"type":"object","properties":{"name":{"type":"string"},"status":{"type":"string"},"background":{"type":"string"},"avatar":{"type":"string"},"sheets":{"$ref":"#/components/schemas/Sheet"},"players":{"$ref":"#/components/schemas/User"},"master":{"$ref":"#/components/schemas/User"},"masterId":{"type":"number"}},"required":["name","status","background","avatar","sheets","players","master","masterId"]},"Sheet":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string"},"image":{"type":"string"},"alignment":{"type":"string"},"background":{"type":"string"},"role":{"type":"string"},"level":{"type":"number"},"experience":{"type":"number"},"health":{"type":"number"},"strength":{"type":"number"},"dexterity":{"type":"number"},"intelligence":{"type":"number"},"constitution":{"type":"number"},"owner":{"$ref":"#/components/schemas/User"},"session":{"$ref":"#/components/schemas/Session"}},"required":["name","gender","image","alignment","background","role","level","experience","health","strength","dexterity","intelligence","constitution","owner","session"]},"UpdateSessionDto":{"type":"object","properties":{"name":{"type":"string"},"background":{"type":"string"},"masterId":{"type":"number"},"avatar":{"type":"string"}}},"CreateSheetDto":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string"},"image":{"type":"string"},"alignment":{"type":"string"},"background":{"type":"string"},"role":{"type":"string"},"level":{"type":"number"},"experience":{"type":"number"},"health":{"type":"number"},"strength":{"type":"number"},"dexterity":{"type":"number"},"intelligence":{"type":"number"},"constitution":{"type":"number"},"ownerId":{"type":"number"},"sessionId":{"type":"number"}},"required":["name","gender","image","alignment","background","role","level","experience","health","strength","dexterity","intelligence","constitution","ownerId","sessionId"]},"UpdateSheetDto":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string"},"image":{"type":"string"},"alignment":{"type":"string"},"background":{"type":"string"},"role":{"type":"string"},"level":{"type":"number"},"experience":{"type":"number"},"health":{"type":"number"},"strength":{"type":"number"},"dexterity":{"type":"number"},"intelligence":{"type":"number"},"constitution":{"type":"number"},"ownerId":{"type":"number"},"sessionId":{"type":"number"}}},"CreateRollDto":{"type":"object","properties":{"expression":{"type":"string"},"ownerId":{"type":"number"}},"required":["expression","ownerId"]},"UpdateRollDto":{"type":"object","properties":{"expression":{"type":"string"},"ownerId":{"type":"number"}}}}}}